clients/layout: move topbar components
